==> code/01_main.R <==
# 01 - Main file for DSP4SH data processing and SOC stock calculation 

# 1 - Run all scripts ----
source(here("code","02_profile_check.R"))
source(here("code","03_missing_msmts.R"))
source(here("code","04_soc_stock_calc.R"))
source(here("code","05_data_compilation.R"))
